backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.1.1
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-serializer-extensions==2.0.1
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
sqlparse==0.4.3
tzdata==2022.4
wrapt==1.14.1
